Engine Fixes (v5/v5u)
This changes:
 - Audio.cpp: fixes incorrect parameters for channel, reflected in GetChannelAttributes for ModAPI as well.
 - Debug.cpp: fixes weird issues when backing out of some menus.
 - Animation.cpp & hpp: removes hitboxCount as it never existed (?), even on early v5 versions.
 - Scene3DLegacyv4: fixes MatrixRotateXYZ rotate functions to match their proper parameters.
   - this is also reflected in RSDKv5 as well, since those were also incorrect.
 - Collision.cpp: Removed workarounds for Mania in v5u mode, implemented needed changes from v5u for CheckObjectCollisionBox (also fixed behaviors for v5 too) & Platform. FindFloorPosition also corrected to match v5 & v5u
 - SceneLegacy (v3, v4): certain variables were actually a part of currentCamera in the official engine, which have been reflected here.
 - PlayerLegacyv3: fixed a chunk of code that didn't exist, also correct some of the P2 follow code.
 - Scene.cpp: v5u fixes for issues with S3K's deformation data (LRZ3 Boss).
   - Scene.hpp also removes the byte difference in id for SceneListEntry.

Minor fixes: Changed Usercore to properly label GamePad mappings. Changed SetupCollisionConfig to match proper parameters.

+++ RSDKv5/RSDK/Audio/Audio.cpp
@@ -506,7 +506,7 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
     return slot;
 }
 
-void RSDK::SetChannelAttributes(uint8 channel, float volume, float panning, float speed)
+void RSDK::SetChannelAttributes(int32 channel, float volume, float panning, float speed)
 {
     if (channel < CHANNEL_COUNT) {
         volume                   = fminf(4.0f, volume);
@@ -524,7 +524,7 @@ void RSDK::SetChannelAttributes(uint8 channel, float volume, float panning, floa
     }
 }
 
-uint32 RSDK::GetChannelPos(uint32 channel)
+uint32 RSDK::GetChannelPos(int32 channel)
 {
     if (channel >= CHANNEL_COUNT)
         return 0;
